+++ packages/core/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @webspatial/core-sdk
 
+## 1.0.1
+
+## 1.0.0
+
+### Major Changes
+
+- bfb72fc: first major version
+
+### Patch Changes
+
+- bc1fcc1: support initScene with resizeRange
+
+## 0.1.23
+
+### Patch Changes
+
+- 3a42a35: Fixed the issue where the default version number of the native was “PACKAGE_VERSION” and the client got the version incorrectly
+
+## 0.1.22
+
+## 0.1.21
+
+## 0.1.20
+
+## 0.1.19
+
+## 0.1.18
+
 ## 0.1.17
 
 ## 0.1.16